requirements.txt
fastapi
uvicorn
hugchat
pydantic
python-dotenv